fix other components of getrow revert

+++ nd4j-api/src/main/java/org/nd4j/linalg/api/buffer/DoubleBuffer.java
@@ -125,7 +125,7 @@ public class DoubleBuffer extends BaseDataBuffer {
 
     @Override
     public void put(int i, int element) {
-        put(i,element);
+        put(i,(double ) element);
     }
